added utility functions mixin
